Tweak upper bounds constraints for dunolint-lib
Remove upper bound constraints directly in the versioned opam file.

So far I was removing these upper bounds at the time of submitting to the public
opam repository. However, for this repo I end up releasing the packages to a
custom opam-repo instead where this extra step is not run.

This is minor but hopefully can help making the intented publication clearer.