Refactor to use qualified path in evaluation
With the introduction of located configs, the path used for evaluation will need
to be the path relative to where the rules where loaded, rather than the path
from the root of the repo. Thus we need to introduce this extra [path] parameter
in order to prepare for the correct passing of qualified path.

This is a refactoring motivated by follow-up changes on config autoloading.